tests(sass): atomic min
