+++ tests/python/test/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_pytest(environment)
-add_pytest(sass)
 
 add_pytest(graph)
 add_pytest(saxpy)
+
+add_subdirectory(sass)
