+++ tests/python/test/sass/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_pytest(atomic)
+add_pytest(sass)
